refactor time
